requirements.txt
python-telegram-bot==13.15
python_dotenv==0.21.1
google-cloud-dialogflow==2.19.1
requests==2.28.2
vk_api==11.9.9
